pyproject.toml
[tool.poetry]
name = "mellitus"
version = "0.1.0"
description = ""
authors = ["Bruno <brunoxbk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
wagtail = "^6.1.2"
django-cors-headers = "^4.4.0"
django-allauth = "^0.63.6"
dj-database-url = "^2.2.0"
dj-rest-auth = "^6.0.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
django-currentuser = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
